fix(e2e): remove scenario cache and fix loading helper
Two compounding issues made canary suites appear to hang after the
[scenario-runner] Cache hit log line:

1. The scenario cache reused provisioned demo companies between local
   runs. For state-mutating tests (any spec that submits a payroll,
   terminates an employee, etc.) cache hits return a company in
   whatever state the previous run left it, breaking repeatability.
   CI never used the cache (no .scenario-cache.json checked in), so
   removing it brings local behavior in line with CI: every test gets
   a fresh demo company. Local re-runs pay the 30-60s provisioning
   cost, which is the honest cost of a repeatable test environment.

2. waitForLoadingComplete polled getByText(/loading/i) and friends,
   matching the SDK's <Loading> Suspense fallback and any per-section
   spinner. It required 3 consecutive non-loading checks and rarely
   got them, silently sitting at 60s timeout. Because Playwright
   does not print step-level progress by default, this manifested as
   "test stalls after Cache hit." Replace with a targeted
   waitFor({ state: 'detached' }) on the Suspense fallback region.

Verified on infrastructure: e2e/tests/payroll.spec.ts now passes all
4 tests in ~12 seconds (vs 3+ minutes per test before the helper fix).
