+++ e2e/scenario/scripts.ts
@@ -1,7 +1,6 @@
 import { resolve } from 'path'
 import { readdirSync, existsSync, readFileSync } from 'fs'
 import { provisionScenario } from './runner'
-import { clearCache } from './cache'
 
 const SCENARIOS_DIR = resolve(process.cwd(), 'e2e/scenarios')
 
@@ -64,12 +63,8 @@ async function main(): Promise<void> {
     case 'prewarm':
       await prewarm(domain)
       break
-    case 'clear':
-      clearCache()
-      console.log('Cache cleared')
-      break
     default:
-      console.error(`Usage: npx tsx e2e/scenario/scripts.ts <prewarm|clear> [--domain <domain>]`)
+      console.error(`Usage: npx tsx e2e/scenario/scripts.ts prewarm [--domain <domain>]`)
       process.exit(1)
   }
 }
