+++ e2e/utils/helpers.ts
@@ -19,44 +19,28 @@ export async function fillDate(
   await dateGroup.getByRole('spinbutton', { name: /year/i }).fill(String(date.year))
 }
 
-export async function waitForLoadingComplete(page: Page, timeout = 60000) {
-  const startTime = Date.now()
-  let consecutiveNonLoadingChecks = 0
-  const requiredConsecutiveChecks = 3
-
-  while (Date.now() - startTime < timeout) {
-    const loadingRegion = page.getByRole('region', { name: /loading/i })
-    const loadingText = page.getByText(/loading/i)
-    const spinner = page.locator('[class*="spinner"], [class*="loading"], [aria-busy="true"]')
-    const skeletonLoader = page.locator('[class*="skeleton"], [class*="placeholder"]')
-
-    const isLoadingVisible =
-      (await loadingRegion.isVisible().catch(() => false)) ||
-      (await loadingText.isVisible().catch(() => false)) ||
-      (await spinner
-        .first()
-        .isVisible()
-        .catch(() => false)) ||
-      (await skeletonLoader
-        .first()
-        .isVisible()
-        .catch(() => false))
-
-    if (!isLoadingVisible) {
-      consecutiveNonLoadingChecks++
-      if (consecutiveNonLoadingChecks >= requiredConsecutiveChecks) {
-        await page.waitForLoadState('networkidle', { timeout: 10000 }).catch(() => {})
-        await page.waitForTimeout(300)
-        return
-      }
-    } else {
-      consecutiveNonLoadingChecks = 0
-    }
-
-    await page.waitForTimeout(200)
-  }
-
-  throw new Error(`Loading did not complete within ${timeout}ms`)
+/**
+ * Wait for the SDK's top-level Suspense fallback (`<Loading>` region with
+ * `aria-label` = `common:status.loading` = "Loading component...") to detach.
+ *
+ * The previous implementation polled four broad selectors including
+ * `getByText(/loading/i)`, which false-matched legitimate page content
+ * (e.g. "Calculating payroll...") and per-section spinners that may
+ * legitimately persist during background polling. It required three
+ * consecutive non-loading checks and rarely got them, silently sitting at
+ * its 60s timeout while Playwright printed no step-level progress.
+ *
+ * Now we wait specifically for the Suspense fallback to detach. Downstream
+ * `expect(...).toBeVisible()` calls handle their own retries, so a stuck
+ * page surfaces as the meaningful landmark-assertion failure rather than a
+ * generic "Loading did not complete."
+ */
+export async function waitForLoadingComplete(page: Page, timeout = 30_000) {
+  const suspenseFallback = page.getByRole('region', { name: /^loading/i })
+  await suspenseFallback
+    .first()
+    .waitFor({ state: 'detached', timeout })
+    .catch(() => {})
 }
 
 export async function skipPendingPayrolls(config: { flowToken: string; companyId: string }) {
